test: game scheduler mock을 testFixtures로 추출
GameSchedulerTestConfig를 :game testFixtures에 정의하고
:admin/:app ServiceTestConfig에서 @Import로 참조.
FlowScheduler/TaskScheduler 추가 시 한 곳만 수정하면 된다.
