+++ backend/admin/build.gradle.kts
@@ -33,6 +33,7 @@ dependencies {
 
     testFixturesImplementation(project(":game-api"))
     testImplementation(project(":test-support"))
+    testImplementation(testFixtures(project(":game")))
     testImplementation(testFixtures(project(":profanity")))
     testImplementation(libs.archunit)
 }
